rename logger.js to Logger.js
